优化 ASTGeneralExpression 关系
