+++ metasequoia_sql/core/node/__init__.py
@@ -2,11 +2,12 @@
 抽象语法树（AST）节点
 """
 
-from metasequoia_sql.core.node.abc_node import ASTBase
-from metasequoia_sql.core.node.basic_node import *
+from metasequoia_sql.core.node.abc_node import ASTBase, ASTExpressionBase
+from metasequoia_sql.core.node.enum_node import *
 from metasequoia_sql.core.node.create_table import *
 from metasequoia_sql.core.node.dql_node import *
 from metasequoia_sql.core.node.insert_node import *
 from metasequoia_sql.core.node.set_node import *
 from metasequoia_sql.core.node.sql_type import SQLType
 from metasequoia_sql.core.node.drop_table import ASTDropTableStatement
+from metasequoia_sql.core.node.basic_node import *
