+++ metasequoia_sql/core/node/abc_node.py
@@ -7,12 +7,12 @@ import dataclasses
 
 from metasequoia_sql.core.node.sql_type import SQLType
 
-__all__ = ["ASTBase"]
+__all__ = ["ASTBase", "ASTExpressionBase"]
 
 
 @dataclasses.dataclass(slots=True, frozen=True, eq=True)
 class ASTBase(abc.ABC):
-    """所有 SQL 语法节点的抽象基类"""
+    """抽象语法树（AST）节点的抽象基类"""
 
     @abc.abstractmethod
     def source(self, sql_type: SQLType = SQLType.DEFAULT) -> str:
@@ -27,3 +27,8 @@ class ASTBase(abc.ABC):
     def get_params_dict(self):
         """获取当前节点的所有参数（用于复制）"""
         return {field.name: getattr(self, field.name) for field in dataclasses.fields(self)}
+
+
+@dataclasses.dataclass(slots=True, frozen=True, eq=True)
+class ASTExpressionBase(ASTBase, abc.ABC):
+    """抽象语法树（AST）一般表达式节点的抽象基类"""
